Major Changes to add new levels
Several changes such as making the players sprite slightly taller in
appearance only so that it looks better overall.  Added in a second
level that already had the collision data made for it. Added a stone
door sprite that coulda looked a lot better if I took more time to make
it :p Added a congratulatory message after completing both levels.